+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of the chatmail cookbook.
 - Debian: Fix mtail service which was not correctly processing logs
 - FreeBSD: Replace the Python filtermail with the new Rust version
 - Suppress an LMTP Received header from being appended to messages (upstream)
+- FreeBSD: Fix package install when version in repo changes
 
 ## 0.2.4